tools/issue-agent/pyproject.toml
[project]
name = "issue-agent"
version = "0.1.0"
description = "GitLab-first issue-to-MR automation workflow"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "python-dotenv>=1.2.1",
  "python-gitlab>=8.0.0",
]

[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/main"]
